requirements.txt
pyopengl
pywavefront
pyglet
Numpy
pygame